modules/ImageClassifierService/requirements.txt
numpy
flask
pillow
tflite-runtime
tflite-support
